Move Secrets types.
