+++ packages/calypso-e2e/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './rest-api-client.types';
 export * from './data-helper.types';
 export * from './editor-tracks.types';
+export * from './secret-manager.types';
